[Classes/Agents] Update state and player to use game models
For the purpose of decision making and using game AI, the game model system GameplayKit has is used. Before, we were using a GameplayKit system for managing states of objects for scenes, which isn't the appropriate system to use.

- AIBaseAgent now is a GKGameModelPlayer
- AIGameState is now a GKGameModel

Some methods in AIGameState still need to be completed for the game model to work.