:sparkles: [Classes/Agents] Add infinite state generation with batch restrictions, clean up code
AI game scenes can now attempt to solve the puzzle infinitely, but with certain restrictions. Attempting to apply updates constantly breaks the game and never lets the window appear to apply those states. So, moves need to be made in batches and put in an SKAction "thread" to ensure that the window will be properly displayed.