:alembic: [Classes/Agents + Structs] Disconnect history tree agent
It turns out that trying to get the history tree agent to make a decision tree based on previous runs is too taxing on this particular project. The code technically works, but a EXC_BAD_ACCESS with code 2 occurs when making the tree because of the fact that it takes too much memory and time to generate the tree, most likely because the perceptrons and neural networks for this get way too complicated for its own good.